bot-telegram/gateway-py/requirements.txt
python-telegram-bot==21.10
httpx==0.27.2
python-dotenv==1.0.1
